restructure db package
