+++ db/db.go
@@ -2,7 +2,6 @@ package db
 
 import (
 	"database/sql"
-	"errors"
 	"os"
 	"regexp"
 )
@@ -112,210 +111,6 @@ func init() {
 	`)
 }
 
-// CreateMessage .
-func CreateMessage(content string, messageGroupID, senderID int) (*Message, error) {
-	result, err := instance.Exec(`
-		INSERT INTO message ( content, message_group_id, sender_id ) VALUES ( ?, ?, ? );
-	`, content, messageGroupID, senderID)
-
-	if err != nil {
-		return nil, errors.New("failed to create message")
-	}
-
-	id, _ := result.LastInsertId()
-	rows, err := instance.Query(`
-		SELECT date_created FROM message WHERE id = ?;
-	`, id)
-
-	if err != nil {
-		return nil, errors.New("failed to find message")
-	}
-
-	if !rows.Next() {
-		return nil, errors.New("can't find message")
-	}
-
-	var dateCreated string
-	rows.Scan(&dateCreated)
-
-	return &Message{
-		ID:             int(id),
-		Content:        content,
-		MessageGroupID: messageGroupID,
-		SenderID:       senderID,
-		DateCreated:    dateCreated,
-	}, nil
-}
-
-// CreateMessageGroup .
-func CreateMessageGroup(groupName string, memberIDs []int) (*MessageGroup, error) {
-	tx, err := instance.Begin()
-	if err != nil {
-		// utils.LogInfo(dbTag, "transaction error. "+err.Error())
-		return nil, err
-	}
-
-	result, err := tx.Exec(
-		`INSERT INTO message_group ( name ) VALUES ( ? );`,
-		groupName,
-	)
-	if err != nil {
-		tx.Rollback()
-		// utils.LogInfo(dbTag, "error creating message group. "+err.Error())
-		return nil, err
-	}
-
-	groupID, err := result.LastInsertId()
-	if err != nil {
-		tx.Rollback()
-		// utils.LogInfo(dbTag, "error creating message group. "+err.Error())
-		return nil, err
-	}
-
-	for _, memberID := range memberIDs {
-		result, err = tx.Exec(
-			`INSERT INTO message_group_member ( message_group_id, member_id )
-			VALUES ( ?, ? );`,
-			groupID,
-			memberID,
-		)
-		if err != nil {
-			tx.Rollback()
-			// utils.LogInfo(
-			// 	dbTag,
-			// 	fmt.Sprintf("failed to create message group %v in instance. %v", groupName, err.Error()),
-			// )
-			return nil, err
-		}
-	}
-
-	err = tx.Commit()
-	if err != nil {
-		// utils.LogInfo(dbTag, "error committing transaction. "+err.Error())
-		return nil, err
-	}
-
-	rows, err := instance.Query(
-		`SELECT date_created FROM message_group WHERE id = ?;`,
-		groupID,
-	)
-	if err != nil {
-		// utils.LogInfo(dbTag, "error retrieving message group. "+err.Error())
-		return nil, err
-	}
-
-	if !rows.Next() {
-		errorMsg := "message_group `%v` has been added to the database but "
-		errorMsg += "an error occured when querying it"
-		// utils.LogInfo(dbTag, fmt.Sprintf(errorMsg, groupName))
-		return nil, errors.New(errorMsg)
-	}
-
-	var dateCreated string
-	rows.Scan(&dateCreated)
-
-	messageGroup := &MessageGroup{
-		ID:          groupID,
-		Name:        groupName,
-		DateCreated: dateCreated,
-
-		MemberIDs: memberIDs,
-	}
-
-	return messageGroup, nil
-}
-
-// CreateMessageGroupMembers .
-func CreateMessageGroupMembers(messageGroupID int, memberIDs []int) error {
-	tx, err := instance.Begin()
-	if err != nil {
-		return err
-	}
-
-	insert, err := tx.Prepare(`
-		INSERT INTO message_group_member ( message_group_id, member_id )
-		VALUES ( ?, ? );
-	`)
-	if err != nil {
-		return err
-	}
-
-	for _, memberID := range memberIDs {
-		_, err := insert.Exec(messageGroupID, memberID)
-		if err != nil {
-			return tx.Rollback()
-		}
-	}
-
-	err = tx.Commit()
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// CreateReceipt .
-func CreateReceipt(messageID, recipientID int) error {
-	_, err := instance.Query(`
-		INSERT INTO receipt ( message_id, recipient_id ) VALUES ( ?, ? );
-	`, messageID, recipientID)
-
-	return err
-}
-
-// DeleteMessageGroupMembers .
-func DeleteMessageGroupMembers(groupID int, memberIDs []int) error {
-	tx, err := instance.Begin()
-	if err != nil {
-		return err
-	}
-
-	delete, err := tx.Prepare(`
-		DELETE FROM message_group_member WHERE message_group_id = ? AND member_id = ?;
-	`)
-	if err != nil {
-		return err
-	}
-
-	for _, memberID := range memberIDs {
-		_, err := delete.Exec(groupID, memberID)
-		if err != nil {
-			return err
-		}
-	}
-
-	err = tx.Commit()
-	if err != nil {
-		tx.Rollback()
-		return err
-	}
-
-	return nil
-}
-
-// FindAllMessageGroupMemberIDs .
-func FindAllMessageGroupMemberIDs(messageGroupID int) ([]int, error) {
-	rows, err := instance.Query(`
-		SELECT member_id
-		FROM message_group_member
-		WHERE message_group_id = ?;
-	`, messageGroupID)
-	if err != nil {
-		return nil, err
-	}
-
-	var memberIDs []int
-
-	for rows.Next() {
-		var id int
-		rows.Scan(&id)
-		memberIDs = append(memberIDs, id)
-	}
-
-	return memberIDs, nil
-}
-
 // FindUndeliveredMessages .
 func FindUndeliveredMessages(recipientID int) ([]*Message, error) {
 	rows, err := instance.Query(`
@@ -346,29 +141,6 @@ func FindUndeliveredMessages(recipientID int) ([]*Message, error) {
 	}
 
 	return messages, nil
-}
-
-// IsClientInMessageGroup .
-func IsClientInMessageGroup(clientID, messageGroupID int) bool {
-	rows, err := instance.Query(`
-		SELECT * FROM message_group_member
-		WHERE message_group_id = ? AND member_id = ?;
-	`, messageGroupID, clientID)
-
-	if err != nil {
-		return false
-	}
-
-	return rows.Next()
-}
-
-// UpdateReceiptDateDelivered .
-func UpdateReceiptDateDelivered(messageID, recipientID int) error {
-	_, err := instance.Exec(`
-		UPDATE receipt SET date_delivered = CURRENT_TIMESTAMP
-		WHERE message_id = ? AND recipient_id = ?;
-	`, messageID, recipientID)
-	return err
 }
 
 func getDigits(s string) string {
